fix included files

+++ build/build-cshared-windows.ps1
@@ -1,8 +1,11 @@
-param ([switch] $RebuildAll, [bool] $IsArm)
+param ([switch] $RebuildAll, [bool] $IsArm = $false, [bool] $CleanBeforeBuild = $true)
 
-Push-Location "$PSScriptRoot\..\src\go"
+Push-Location "$PSScriptRoot/../src/go"
 
-Remove-Item -r -fo bin -ea SilentlyContinue
+if($CleanBeforeBuild)
+{
+	Remove-Item -r -fo bin -ea SilentlyContinue
+}
 
 $lib_name = "ProtonSecurity"
 
@@ -11,10 +14,11 @@ $env:CGO_ENABLED = "1"
 $env:CGO_LDFLAGS = "-s -w"
 $env:GOOS = "windows"
 
-$env:GOARCH = If($IsArm) { "arm64" } Else { "amd64" }
-$pathArch = If($IsArm) { "win-arm64" } Else { "win-x64" }
+$env:GOARCH = if($IsArm) { "arm64" } else { "amd64" }
+$pathArch = if($IsArm) { "win-arm64" } else { "win-x64" }
 
-$outputPath = "bin\runtimes\$pathArch\native\$lib_name.dll"
+$outputPath = "/bin/runtimes/$pathArch/native/$lib_name.dll"
+write-host "Building to `"$(get-location)$outputPath`""
 $buildCommand = "go build -buildmode=c-shared -v -o `"$outputPath`""
 if ($RebuildAll) {
 	$buildCommand += " -a"
